fix bug update crop ảnh https://jira.amitech.vn/browse/VDLGS-64 https://jira.amitech.vn/browse/VDLGS-19

+++ CMS/Areas/Orders/Models/CreateOrderModel.cs
@@ -6,9 +6,9 @@ namespace CMS.Areas.Orders.Models;
 public class CreateOrderModel
 {
     public List<ProductCheckoutViewmodel>? Products { get; set; }
-    public int? PriceShip { get; set; }
-    public int? PriceNoSale { get; set; }
-    public int? Percent { get; set; }
+    public double? PriceShip { get; set; }
+    public double? PriceNoSale { get; set; }
+    public double? Percent { get; set; }
     public double? Total { get; set; }
     public int? CouponDiscount { get; set; }
     public int? CustomerId { get; set; }
